AttackerComboTest2
Using values of second SPSA session "as is".
Some values are strange, I guess the accuracy is +-5. Also, infrequent
combo can have any value, that will not change anything. Finally, when
we compare some 2 combos such as rr with 3 combos such as rrb or rrn,
the values are surprising, but keep in mind that as the attacks builds
up, the defense builds up too, so we cannot really compare rr with rrb
or rrn.