FCC_V1
Faster nps than master, and as a bonus, more accurate:
-Safe contact checks by Queens are now computing the checks supported by
King or another Queen.
-Safe contact checks by Rooks are now computing the checks supported by
King or other Rook

This make quite a change, (especially in the common RR case) so if test
fails, will retry by adjusting special cases QQ contact check value and
RR contact check value. (which were valued 0 before)

For example. Black Rf8, Re1. White Nf1 and Kg1.
Black has now 2 safe contact checks which were not recorded before, that
is 68 * 2 more added to the attackUnits !

For example. Black Rf8, Rf2. White Nf1 and Kg1.
Black has now 1 contact check which was not recorded before, that is 68
more. (x-rays are handled properly).

For the QQ case, the change is 92 or 194.