EPS_V1
When a Minor is attacked by a pawn, it does not make much sense to
proceed to calculate the outpost value (it will be 0).  In this bold
attempt, we skip also other Minor bonus/penalties.

The rationale is a small speed up:
Results for 5 tests for each version:

Base      Test      Diff
Mean    1355576   1365439   -9863
StDev   70301     68833     3086

p-value: 0.999