MP_20150323_V1
Using a new global pawn span (the column distance between far most pawns
of any side) which is a number between 0 and 7.

Give bonus for Bishop when global pawn span is >5
and to equilibrate things, a bonus for Knight when global pawn span is <
5.

Different then snicolet recent test and binky july 2014 tests since
a) it is a one-time bonus, not increased by # column
b) it is never a penalty (because we do not care about the exact piece #
and configuration)
c) using a S(0,12) aiming for endgame.
d) similar type of bonus for both Minors, so we are not skewing only one
type of piece
e) and using a global pawn span instead or pawn span of on or the other
color.

Freely adapted based on ideas and discussions with Fauzi.

Benc # 7862627