RM_20150901-1
We currently remove squares controlled by lesser pieces in the queen
mobility count.
Doing the same for the rook was tested some time ago,
and did not worked,
In this test, we remove squares controlled by lesser
pieces only when the rook is working alone on those squares, or not
supporting/defending anything.

A typical case is when White Ra1, Black Bb7, Bb6
If a5, a6, a7 and a8
are attacked by the White rook alone,
this test does not considered the
Rook mobile or "useful" on those squares.