SRA_20151212_1
I have try anything under the sun with the RookAligned bonus...trying to
improve it or remove undesirable side effects, tuning by row, and so on.

I'm aware also of a test by Lucas where he tried a simplification by
adjusting the rook pst table to somplensate the removal of the bonus.

But recently, a test intending an adjustment to "rook Aligned" bonus was
in fact cancelling it (by mistake)
http://tests.stockfishchess.org/tests/view/5664c7720ebc5902cdc08993

The test was stopped due to mistale,
but the result after 1300 games
LLR: -0.16 (-2.94,2.94) [0.00,5.00]
Total: 1313 W: 257 L: 258 D: 798

This shows that the bonus can be removed.

Bench:9605822
Master: 9264977