UA_20160109_1
Give a bonus instead for coordination:
a Minor and a Pawn attack a pawn-defended pawn.

As soon as the pawn is exchanged, the Threat[Minor][PAWN] will kick in.

So this is a new MinorOnPawn bonus is something in between , and will
act as a "keep the tension" type of bonus.

bench: 8079715
master:8603285