RIF_20160308_1
Currently we add mobile pawns on rank 4 and above in the mobiityarea.
This is perfect for most pieces. Indeed, if the pawn push, this increase
our mobility.

But there is one case where it is not good at all.
Q and R when in front of that mobile pawn

Example: White e4 and Re6: e4 should not be part of the mobility area of
R since a pawn push will not increase the rook mobility, it will reduce
it !

Let's see if this will make any difference.
bench: 8007881