SP_20160920
SEE is used for give some score for an initial move,  considering the
safety of the landing square and the gradual exchanges on that
square.

The aim of this test is to penalize longer see sequences.
The
longer a sequence, the less likely it will be played, (attackers can be
captured on other squares, in between moves, pieces that are not pinned
but have crucial defensive task, and so on).

bench: 5547519
master: 6116200