B2_20161121_1
Consider a position where White have pawns on b2 and a3.

b2 is currently "backward" if black has a pawn on a4, b3 and/or c4

This test gives an extra penalty when case 2 and 3 does not apply,
because a single pawn (black a4) is functionally holding 2 pawns.

bench: 6318493
master: 5773672