SSC_20170107_1
Another check experiment
More bonus for safe Knight and Rook checks given on a supported square,
because if king comes closer, it can attack the knight or the rook (at
distance 2), but not a bishop or a queen.
More generally supported rooks are good if user interpose with a rook,
or is a possible sign that another rook is not far away.

bench: 5322885
master: 4940355