SO_20160116_1
Introducing smart outposts for Knights
If a pawn tries to kick the Knight, it creates another outpost on the
next rank.
Example: White Ne5, Black f7 g7 h6
Ne5 is not a traditional outpost since it can be kicked by the f pawn.
But since g6 is only defended once, if f7-f6 then g6 is a new reachable
outpost for the knight.

bench: 5805489
master: 5367704