QS_20170318
Add squares supported by queen through other sliders.
For example Qh1 Rh2 Rg1 and Bg2
With this "fix", the squares h3 to h8, f1 to a1 and f3 to a8 are added
to the attackedBy2 bitboard.

This create more cases where we consider a square safe for checking,
higher closeenemies, and more "strongly defended" when evaluating
threats.

bench: 6414321
master: 6128779