TRC_20170918
Tying to improve ThreatByRank.
Somehow the current logic is screwed.
If I'm white, and black pieces are all on the 8th rank,
white will get some base bonus for each of his threats,
but is total ThreatByRank  will be 0.

However, if black as a single threat on my Ra8, he gets a S(16,3 )* 7
additional bonus.

Trying to fix this, we locate our closest enemy, say Nd7,
its rdelta is 6, so White ThreatByRank factor will be 7 for his attacks
on row 7, and 6 for his attacks on row 8.

Maybe a better formula would be suitable for rdelta (such as divide
rdelta by 2)

bench 6106065
master 6259498