SLD2_201710
Current definition allows our bishop to attack one pawn in the center.
So a bishop in the center will always have the S(22, 0) bonus

Let's move this to the psqt.
Bench would be the same as master if it was not from the LazyEval which
will rarely kicks in in the midgame.

bench: 5126909
master:  5234652