CB2_20180122_2
White d4, e5, Black e6

Master was considering e5 as "passed" (with half bonus) (because d4-d5
would make the way for e5)

On a board with only pawns, this is indeed correct.
But there can be a piece on d5 blocking the d4 pawn,
or the d5 square might be sufficiently protected.

So this version
a) give the half passer bonus only if there is no blocker on d5
b) give the half passer bonus also if we have a pawn on d5 or f5

bench: 5085853
master: 4971497