BPM_20180220
Bishop pair bonus is 1667 in the imbalance table of material.cpp. This
translate to S(104, 104)
Compare this to a pawn value which is S(171, 240)

Master is currently assigning different mobility value for each bishop.

This test tries to "also" assign a mobility bonus /penalty by looking at
how the 2 bishops works together.

I use the sum of the mobility of both bishop, and if below 10, there
will be a larger penalty. If above 10, a smaller bonus.

Another measure could simply be to find which bishop is the strongest of
the 2, and adjust the bishop pair accordingly.

bench 5084244
master 5006365