CS1_b
bench 3996491
compute a few more "unsafe checks"
for example
A) White Bc4 and Black Ke8 and pawn f7
master computed Bb5 as a safe checks, and does not score any unsafe
checks for bishop.

B) White Bc4 and Black Ke8 and pawn f7 and a6
master computes all the unsafe bishop checks, Bb5 and Bf7.

This version also gives a base bonus 148 for the Bxf7 in case A)